refactor: sessions 패키지 분리
