+++ src/main/java/nextstep/sessions/domain/enrollment/Capacity.java
@@ -1,4 +1,4 @@
-package nextstep.sessions.domain;
+package nextstep.sessions.domain.enrollment;
 
 public class Capacity {
