+++ src/main/java/nextstep/sessions/domain/image/FileName.java
@@ -1,4 +1,4 @@
-package nextstep.sessions.domain;
+package nextstep.sessions.domain.image;
 
 public class FileName {
