+++ src/main/java/nextstep/sessions/domain/image/ImageType.java
@@ -1,4 +1,4 @@
-package nextstep.sessions.domain;
+package nextstep.sessions.domain.image;
 
 import java.util.Arrays;
