+++ src/test/java/nextstep/sessions/domain/enrollment/EnrollmentTest.java
@@ -1,4 +1,4 @@
-package nextstep.sessions.domain;
+package nextstep.sessions.domain.enrollment;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
