+++ src/test/java/nextstep/sessions/domain/image/ImageDimensionTest.java
@@ -1,4 +1,4 @@
-package nextstep.sessions.domain;
+package nextstep.sessions.domain.image;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
